reorganized charts and deleted extraneous ones
