--- src/components/Services/ApacheMetrics/Charts.tsx
@@ -1,3 +1,0 @@
-export default function Chart() {
-  return <h1>I am a Apache Metric chart!</h1>;
-}
