--- src/components/Services/CustomApplication/Charts.tsx
@@ -1,3 +1,0 @@
-export default function Chart() {
-  return <h1>I am a Custom App chart!</h1>;
-}
