--- src/components/Services/Error.tsx
@@ -1,3 +1,0 @@
-export default function Error() {
-  return <h1>Error: Service not found</h1>;
-}
