--- src/components/Services/MongoLogs/Charts.tsx
@@ -1,3 +1,0 @@
-export default function Chart() {
-  return <h1>I am a Mongo Logs chart!</h1>;
-}
