--- src/components/Services/NginxErrorLogs/Charts.tsx
@@ -1,3 +1,0 @@
-export default function Chart() {
-  return <h1>I am a Nginx Log chart!</h1>;
-}
