--- src/components/Services/PostgresLogs/Charts.tsx
@@ -1,3 +1,0 @@
-export default function Chart() {
-  return <h1>I am a Postgres Log chart!</h1>;
-}
